+++ lib/glimmer/tk/button_proxy.rb
@@ -20,6 +20,7 @@
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 require 'glimmer/tk/widget_proxy'
+require 'glimmer/tk/commandable'
 
 module Glimmer
   module Tk
@@ -27,20 +28,7 @@ module Glimmer
     #
     # Follows the Proxy Design Pattern
     class ButtonProxy < WidgetProxy
-      # TODO extract to a module
-      
-      def command_block=(proc)
-        tk.command(proc)
-      end
-      
-      def handle_listener(listener_name, &listener)
-        case listener_name.to_s.downcase
-        when 'command'
-          command(listener)
-        else
-          super
-        end
-      end
+      include Commandable
     end
   end
 end